% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make up missing values.R
\name{make up missing values}
\alias{make up missing values}
\alias{MakeupMissingValue}
\title{Make up missing values}
\usage{
MakeupMissingValue(data)
}
\arguments{
\item{data}{the matrix may be has missing values}
}
\value{
a matrix has no missing values
}
\description{
Make up missing values by median
}
\examples{
\dontrun{
x<-rnorm(20)
data<-rnorm(x,4,5)
data[2,3]<-NA
MakeupMissingValue(data)
}
}
