% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwC}
\alias{rwC}
\title{A Metropolis sampler using Rcpp}
\usage{
rwC(sigma, x0, N)
}
\arguments{
\item{sigma}{variance of the proposal distribution}

\item{x0}{the initial value}

\item{N}{the number of samples}
}
\value{
a randpm sample of Cauchy distribution and the number of candidate points rejected
}
\description{
A Metropolis sampler using Rcpp
}
\examples{
\dontrun{
dir_cpp <- 'D:/Rstudio/Rcpp/'
source(paste0(dir_cpp,'rwR.R'))
sourceCpp(paste0(dir_cpp,'rwC.cpp'))
N <- 2000
sigma <- .5
x0 <- 25
rwC <- rwC(sigma, x0, N)
print(rwC[N+1]/N)
}
}
