\name{SC19060-package}
\alias{SC19060-package}
\alias{SC19060}
\docType{package}
\title{
  Make up missing values and Metropolis Sampler.
}
\description{
  Make up missing values by median and generate the Cauchy distribution by Metropolis sampler.
}
\author{
Minhui Zhu

Maintainer: Minhui Zhu <zminhuii@163.com>
}
\keyword{ package }
